--- src/main/java/raisetech/studentManagement/StudentRepository.java
@@ -1,15 +1,0 @@
-package raisetech.studentManagement;
-
-import java.util.List;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Select;
-
-@Mapper
-public interface StudentRepository {
-
-  @Select("SELECT * FROM students")
-  List<Student> searchStudents();
-
-  @Select("SELECT * FROM students_courses")
-  List<StudentCourse> searchStudentCourse();
-}
