+++ src/main/java/raisetech/studentManagement/repository/StudentRepository.java
@@ -1,0 +1,28 @@
+package raisetech.studentManagement.repository;
+
+import java.util.List;
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Select;
+import raisetech.studentManagement.date.Student;
+import raisetech.studentManagement.date.StudentCourse;
+
+/**
+ * 受講生情報を扱うリポジトリ。
+ *
+ * 全件検索や単一条件での検索、コース情報の検索が行えるクラスです。
+ */
+@Mapper
+public interface StudentRepository {
+
+  /**
+   * 全件検索します。
+   *
+   * @return　全件検索した受講生情報の一覧
+   */
+
+  @Select("SELECT * FROM students")
+  List<Student> searchStudents();
+
+  @Select("SELECT * FROM students_courses")
+  List<StudentCourse> searchStudentCourse();
+}
